Updated Keithley SM after trying to instrument
The instrument is not wanting to cooperate with me. It appears that the
GET functions (of what I've written) are working; however, the set
functions are not even executed when I hit set in the GUI. I'll have to
have a bit more of a play around with the language.