Lab Brick Driver
Driver for a Lab Brick frequency synthesizer device. Tested with a
LMS-152D under Scientific Linux and Windows 7. The driver is still being
developed so functionality is limited. There is a big problem with
threading in the driver. The library of the Lab Brick uses a thread to
read out the response from the instrument. When this library is reloaded
(which some of us do every time a measurement script is run) the new
instance can not read out the response anymore. If you want to use this
driver, load it only once.
This driver depends on a library made by Vaunix. For Linux they provide
the source code of the library, for Windows I only have a compiled
library but I haven't asked for the source. Because of this I don't
think it's completely GPL compliant.