New wrapper for SIM900 mainframe
An alternative for the separate wrappers for the modules, this wrapper
provides one instrument that detects what modules are present and
creates functions to control those modules. A benefit is that there is
only one wrapper communicating with the SIM900 mainframe, instead of
several wrappers sharing the same VISA communication channel.